create: create screens by using StackNavigator
