+++ App.js
@@ -1,46 +1,8 @@
-import React, {useState, useEffect} from 'react';
-import { StyleSheet, FlatList, SafeAreaView } from 'react-native';
-import ListItem from './components/ListItem'
-import dummyArticles from './dummies/articles'
-import Constants from 'expo-constants'
-import axios from 'axios'
-
-const URL = `http://newsapi.org/v2/top-headlines?country=jp&category=business&apiKey=${Constants.manifest.extra.newsApiKey}`
+import React from 'react';
+import AppNavigator from "./navigation/AppNavigator"
 
 export default function App() {
-  const [articles, setArticles] = useState([])
-  useEffect(() => {
-    fetchArticles();
-  }, []);
-
-  const fetchArticles = async() => {
-    try {
-      const response = await axios.get(URL)
-      setArticles(response.data.articles)
-    } catch (error) {
-      console.log(error)
-    }
-  };
   return (
-    <SafeAreaView style={styles.container}>
-      <FlatList
-        data={articles}
-        renderItem={({item}) => (
-          <ListItem
-            imageUrl={item.urlToImage}
-            title={item.title}
-            author={item.author}
-          />
-        )}
-        keyExtractor={(item, index) => index.toString()}
-      />
-    </SafeAreaView>
+    <AppNavigator />
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff'
-  }
-});
